backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
pydantic[email]==2.6.3
pydantic-settings==2.1.0
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
minio==7.2.0
celery==5.3.6
redis==5.0.1
pillow==10.2.0
numpy==1.26.4
scikit-image==0.22.0
python-dotenv==1.0.1
bcrypt==4.1.2
opencv-python==4.8.1.78
torch==2.2.1
albumentations==1.3.1
torchvision==0.17.1
psutil==5.9.8
boto3==1.34.69